# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SiapPA API",
        "description": "API documentation for SiapPA application",
        "contact": {
            "email": "support@siappa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/pengaduans": {
            "get": {
                "tags": [
                    "Pengaduans"
                ],
                "summary": "Get list of pengaduans",
                "operationId": "9550c8ea3f98239b985e942b0211473d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pengaduans"
                ],
                "summary": "Create a new pengaduan",
                "operationId": "04b58eb148343c76f2a8a761e4ada2de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PengaduanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/pengaduans/{id}": {
            "get": {
                "tags": [
                    "Pengaduans"
                ],
                "summary": "Get a specific pengaduan",
                "operationId": "7789046da6d1a9646dfc5db8fdc5510e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pengaduans"
                ],
                "summary": "Update a specific pengaduan",
                "operationId": "1bb482b8cb0ecbecef903ba61e84f86c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PengaduanRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pengaduans"
                ],
                "summary": "Delete a pengaduan",
                "operationId": "ccd82abaa6ebecfcfcc29de2ef027a3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "PengaduanRequest": {
                "required": [
                    "namaKorban",
                    "alamat",
                    "aduan",
                    "kategoriKekerasan",
                    "harapan"
                ],
                "properties": {
                    "namaKorban": {
                        "type": "string",
                        "example": "Siti Nurhaliza"
                    },
                    "alamat": {
                        "type": "string",
                        "example": "Jl. Merdeka No. 123"
                    },
                    "aduan": {
                        "type": "string",
                        "example": "Saya mengalami kekerasan."
                    },
                    "kategoriKekerasan": {
                        "type": "string",
                        "enum": [
                            "kekerasan_fisik",
                            "kekerasan_seksual",
                            "kekerasan_lainnya"
                        ],
                        "example": "kekerasan_fisik"
                    },
                    "harapan": {
                        "type": "string",
                        "example": "Saya berharap pelaku dihukum."
                    },
                    "status": {
                        "type": "string",
                        "example": "terkirim",
                        "nullable": true
                    },
                    "evidenceUrls": {
                        "type": "string",
                        "example": "https://example.com/evidence.jpg",
                        "nullable": true
                    },
                    "evidencePaths": {
                        "type": "string",
                        "example": "evidence/image.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}